+++ src/docs/stories/13-CardList.stories.js
@@ -7,6 +7,7 @@ import ListItem from '../../components/atom/list-item/ListItem';
 import List from '../../components/molecule/list/List';
 import Card from '../../components/molecule/card/Card';
 import CardItem from '../../components/atom/card-item/CardItem';
+import EmptyCardItem from '../../components/atom/empty-card-item/EmptyCardItem';
 
 import Wrapper from '../../helpers/wrapper/Wrapper';
 import CardLine from '../../components/organism/card-line/CardLine';
@@ -17,7 +18,7 @@ export default {
   component: Card
 };
 
-export const CardList = () => (
+export const Default = () => (
   <Wrapper style={{ maxWidth: 800, marginTop: 10 }}>
     <List>
       <ListItem>
@@ -53,41 +54,51 @@ export const small = () => (
   </Wrapper>
 );
 
-export const mLg = () => (
-  <Wrapper>
-    <List level={4}>
+export const WithCheckBox = () => (
+  <Wrapper style={{ maxWidth: '100%', marginTop: 10 }}>
+    <List>
       <ListItem>
         <Card>
-          <CardItem value="some value" label="label" />
-        </Card>
-      </ListItem>
-      <ListItem>
-        <Card>
-          <CardItem value={new Date()} label="Última atualização" />
+          <EmptyCardItem>
+            <Input
+              id="ipt"
+              name="ipt"
+              type="checkbox"
+              defaultChecked
+              value="1"
+              onChange={action('toggleCheck')}
+            />
+          </EmptyCardItem>
+          <CardItem value="1" label="#" />
+          <CardItem value="Ativo" label="Status" />
+          <CardItem value="Matheus Alexandre" label="Nome do orçamento" />
+          <CardItem value="04 de Set 2019 às 13:58" label="Criado em" />
+          <CardItem value="04 de Set 2019 às 13:58" label="Atualizado em" />
         </Card>
       </ListItem>
     </List>
   </Wrapper>
 );
 
-export const mSm = () => (
-  <Wrapper>
-    <List level={1}>
+export const CardListActionSmall = () => (
+  <Wrapper style={{ maxWidth: 800, marginTop: 10 }}>
+    <List>
       <ListItem>
-        <Card>
-          <CardItem value="Some value" label="Label" primary />
-        </Card>
-      </ListItem>
-      <ListItem>
-        <Card>
-          <CardItem value="Value" label="Label" />
-          <CardItem value="Value2" label="Label" />
-        </Card>
-      </ListItem>
-      <ListItem>
-        <Card>
-          <CardItem value="Value" label="Label" col={8} />
-          <CardItem value="Value2" label="Label" col={4} />
+        <Card small>
+          <EmptyCardItem small>
+            <Input
+              id="ipt"
+              name="ipt"
+              type="checkbox"
+              defaultChecked
+              value="1"
+              onChange={action('toggleCheck')}
+            />
+          </EmptyCardItem>
+          <CardItem value="Ativo" label="Status" small />
+          <CardItem value="Matheus Alexandre" label="Nome do orçamento" small />
+          <CardItem value="04 de Set 2019 às 13:58" label="Criado em" small />
+          <CardItem value="04 de Set 2019 às 13:58" label="Atualizado em" small />
         </Card>
       </ListItem>
     </List>
@@ -117,14 +128,16 @@ export const _CardLine = () => (
     <List level={5}>
       <ListItem>
         <CardLine>
-          <Input
-            id="ipt"
-            name="ipt"
-            type="checkbox"
-            defaultChecked
-            value="1"
-            onChange={action('toggleCheck')}
-          />
+          <EmptyCardItem>
+            <Input
+              id="ipt"
+              name="ipt"
+              type="checkbox"
+              defaultChecked
+              value="1"
+              onChange={action('toggleCheck')}
+            />
+          </EmptyCardItem>
           <CardItem
             value="Value"
             label="Label"