+++ src/index.js
@@ -13,6 +13,7 @@ import BoxShadow from './components/atom/box-shadow/BoxShadow';
 import Button from './components/atom/button/Button';
 import CardItem from './components/atom/card-item/CardItem';
 import DatePicker from './components/atom/datepicker/DatePicker';
+import EmptyCardItem from './components/atom/empty-card-item/EmptyCardItem';
 import FileInput from './components/atom/file-input/FileInput';
 import Heading from './components/atom/heading/Heading';
 import Input from './components/atom/input/Input';
@@ -35,7 +36,6 @@ import FormActions from './components/molecule/form-actions/FormActions';
 import FormGroup from './components/molecule/form-group/FormGroup';
 import FormRow from './components/molecule/form-row/FormRow';
 import List from './components/molecule/list/List';
-import ListItemContent from './components/molecule/list-item-content/ListItemContent';
 import ModalContent from './components/molecule/modal-content/ModalContent';
 import ModalFooter from './components/molecule/modal-footer/ModalFooter';
 import ModalHeader from './components/molecule/modal-header/ModalHeader';
@@ -49,7 +49,6 @@ import CardLine from './components/organism/card-line/CardLine';
 import Form from './components/organism/form/Form';
 import FormModal from './components/organism/form-modal/FormModal';
 import Filter from './components/organism/filter/Filter';
-import ListItemCWrapper from './components/organism/list-item-c-wrapper/ListItemCWrapper';
 import Modal from './components/organism/modal/Modal';
 
 export {
@@ -63,6 +62,7 @@ export {
   DatePicker,
   Dropdown,
   DropdownMenu,
+  EmptyCardItem,
   FileInput,
   Filter,
   FilterHeader,
@@ -79,8 +79,6 @@ export {
   Label,
   List,
   ListItem,
-  ListItemContent,
-  ListItemCWrapper,
   Loader,
   MaskedInput,
   Modal,